requirements.txt
# Flask web framework
flask==3.0.3

# Flask CORS support
flask-cors==5.0.0

# Swagger/OpenAPI documentation
flasgger==0.9.7.1

# OpenTelemetry core
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0

# OTLP exporter (gRPC)
opentelemetry-exporter-otlp-proto-grpc==1.27.0

# Flask auto-instrumentation
opentelemetry-instrumentation-flask==0.48b0

# PostgreSQL driver
psycopg2-binary==2.9.10

# PostgreSQL auto-instrumentation
opentelemetry-instrumentation-psycopg2==0.48b0

# Production WSGI server
gunicorn==23.0.0

# Environment variable management
python-dotenv==1.0.1

# gRPC client for otel-worker integration
grpcio==1.62.0
grpcio-tools==1.62.0
protobuf==4.25.3

# gRPC auto-instrumentation
opentelemetry-instrumentation-grpc==0.48b0
